feat: Create an 'examples' project with full use-cases
The examples project should be a collection of complete, yet simple,
use-cases that would allow the user to have a better understanding about
capabilities offered by funktions and how to get started.

Furthermore, the README.md was updated to provide a better overview
and a link to the examples directory.